Fix intermittant crashes in location
Deleting memory without setting it to null
can cause the memory to still be non-null
and be interpreted as valid.
Also check if the entry is in the map using
find instead of assuming it is in the map.

Bug: 62033719
Bug: 62033834
Bug: 62033690
Bug: 62033563
Bug: 62032790
CRs-fixed: 2050837
Change-Id: I2534de2d6157be86ac95cfe6615e4c0019ee48fd
(cherry picked from commit 5a81c9b089f1a35501863aa1bf738029e930d05a)